refactor(audio): extract AudioDeviceManager from main.cpp
