+++ Audio/AudioDeviceManager.h
@@ -1,0 +1,28 @@
+#pragma once
+
+#include "../miniaudio.h"
+
+class AudioDeviceManager
+{
+  public:
+    ~AudioDeviceManager();
+
+    bool initializeCapture();
+    bool initializePlayback();
+    bool startDevices();
+    bool startCapture();
+    bool startPlayback();
+    bool restartPlayback();
+    void stopDevices();
+
+    static void capture_callback(ma_device *device, void *output, const void *input, ma_uint32 frameCount);
+    static void playback_callback(ma_device *device, void *output, const void *input, ma_uint32 frameCount);
+
+  private:
+    ma_device captureDevice{};
+    ma_device playbackDevice{};
+    bool captureInitialized = false;
+    bool playbackInitialized = false;
+
+    void cleanup();
+};
