+++ Resona.vcxproj
@@ -134,12 +134,14 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="Audio\AudioDeviceManager.cpp" />
     <ClCompile Include="GpuUpsampler.h" />
     <ClCompile Include="main.cpp" />
     <ClCompile Include="RenderPipeline\SubmissionPlanner.cpp" />
     <ClCompile Include="VulkanUpsampler.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="Audio\AudioDeviceManager.h" />
     <ClInclude Include="miniaudio.h" />
     <ClInclude Include="RenderPipeline\SubmissionPlanner.h" />
     <ClInclude Include="VulkanUpsampler.h" />
