+++ main.cpp
@@ -27,6 +27,7 @@
 #define MINIAUDIO_IMPLEMENTATION
 #include "miniaudio.h"
 
+#include "Audio/AudioDeviceManager.h"
 #include "GpuUpsampler.h"
 #include "RenderPipeline/SubmissionPlanner.h"
 #include "VulkanUpsampler.h"
@@ -378,234 +379,6 @@ void signal_handler(int signal)
         g_running = false;
     }
 }
-
-// === RAII Audio Device Manager ===
-class AudioDeviceManager
-{
-  public:
-    ~AudioDeviceManager()
-    {
-        cleanup();
-    }
-
-    bool initializeCapture()
-    {
-        ma_device_config config = ma_device_config_init(ma_device_type_capture);
-        config.sampleRate = AudioConfig::INPUT_SAMPLE_RATE;
-        config.capture.format = ma_format_f32;
-        config.capture.channels = AudioConfig::CHANNELS;
-        config.dataCallback = capture_callback;
-
-        // === Performance Optimizations ===
-
-        // Disable miniaudio resampling and conversion (already done, keep these)
-        config.noPreSilencedOutputBuffer = MA_TRUE;
-        config.noClip = MA_TRUE;
-        config.noFixedSizedCallback = MA_TRUE;
-
-        // NEW: Set performance profile for low latency
-        config.performanceProfile = ma_performance_profile_low_latency;
-
-// NEW: WASAPI-specific optimizations (Windows only)
-#ifdef MA_WIN32
-        config.wasapi.noAutoConvertSRC = MA_TRUE;      // Disable WASAPI's sample rate converter
-        config.wasapi.noDefaultQualitySRC = MA_TRUE;   // Disable quality SRC for speed
-        config.wasapi.noHardwareOffloading = MA_FALSE; // Allow hardware offloading
-#endif
-
-        if (ma_device_init(nullptr, &config, &captureDevice) != MA_SUCCESS)
-        {
-            printf("[!] Failed to initialize capture device\n");
-            return false;
-        }
-
-// Set high priority for capture thread (Windows)
-#ifdef _WIN32
-        std::this_thread::sleep_for(std::chrono::milliseconds(10));
-
-        HANDLE threadHandle = (HANDLE)captureDevice.thread;
-        if (threadHandle && threadHandle != INVALID_HANDLE_VALUE)
-        {
-            if (!SetThreadPriority(threadHandle, THREAD_PRIORITY_TIME_CRITICAL))
-            {
-                printf("[*] Warning: Could not set capture thread priority (error: %lu)\n", GetLastError());
-            }
-            else
-            {
-                printf("[+] Capture thread priority set to TIME_CRITICAL\n");
-            }
-        }
-#endif
-
-        // Report actual device configuration
-        printf("[+] Capture device initialized: %u Hz, %u frames/period, %u periods\n", captureDevice.sampleRate,
-               captureDevice.capture.internalPeriodSizeInFrames, captureDevice.capture.internalPeriods);
-
-        // Calculate and report callback frequency
-        const float callbackFrequency =
-            static_cast<float>(captureDevice.sampleRate) / captureDevice.capture.internalPeriodSizeInFrames;
-        const float callbackIntervalMs = 1000.0f / callbackFrequency;
-
-        printf("[+] Capture callback: %.1f Hz (every %.2f ms)\n", callbackFrequency, callbackIntervalMs);
-
-        captureInitialized = true;
-        return true;
-    }
-
-    bool initializePlayback()
-    {
-        ma_device_config config = ma_device_config_init(ma_device_type_playback);
-        config.sampleRate = AudioConfig::OUTPUT_SAMPLE_RATE;
-        config.playback.format = ma_format_f32;
-        config.playback.channels = AudioConfig::CHANNELS;
-        config.dataCallback = playback_callback;
-
-        // === Performance Optimizations ===
-
-        // Disable ALL miniaudio resampling and conversion
-        config.noPreSilencedOutputBuffer = MA_TRUE;
-        config.noClip = MA_TRUE;
-        config.noFixedSizedCallback = MA_TRUE; // Allow variable-size callbacks
-
-        // NEW: Set performance profile for low latency
-        config.performanceProfile = ma_performance_profile_low_latency;
-
-// NEW: WASAPI-specific optimizations (Windows only)
-#ifdef MA_WIN32
-        config.wasapi.noAutoConvertSRC = MA_TRUE;      // Disable WASAPI's sample rate converter
-        config.wasapi.noDefaultQualitySRC = MA_TRUE;   // Disable quality SRC for speed
-        config.wasapi.noHardwareOffloading = MA_FALSE; // Allow hardware offloading
-
-        // NEW: Try to use exclusive mode for lowest latency (may fail, that's OK)
-        config.wasapi.usage = ma_wasapi_usage_pro_audio; // Pro audio usage hint
-#endif
-
-        if (ma_device_init(nullptr, &config, &playbackDevice) != MA_SUCCESS)
-        {
-            printf("[!] Failed to initialize playback device\n");
-            return false;
-        }
-
-        // Report actual device sample rate
-        printf("[+] Playback device initialized: %u Hz (requested: %u Hz), %u frames/period, %u periods\n",
-               playbackDevice.sampleRate, AudioConfig::OUTPUT_SAMPLE_RATE,
-               playbackDevice.playback.internalPeriodSizeInFrames, playbackDevice.playback.internalPeriods);
-
-        playbackInitialized = true;
-        return true;
-    }
-
-    bool startDevices()
-    {
-        if (!captureInitialized || !playbackInitialized)
-        {
-            return false;
-        }
-
-        // Start capture and playback together
-        if (ma_device_start(&captureDevice) != MA_SUCCESS)
-        {
-            printf("[!] Failed to start capture device\n");
-            return false;
-        }
-
-        // Start playback device
-        if (ma_device_start(&playbackDevice) != MA_SUCCESS)
-        {
-            printf("[!] Failed to start playback device\n");
-            ma_device_stop(&captureDevice);
-            return false;
-        }
-
-        return true;
-    }
-
-    bool startCapture()
-    {
-        if (!captureInitialized)
-        {
-            return false;
-        }
-
-        if (ma_device_start(&captureDevice) != MA_SUCCESS)
-        {
-            printf("[!] Failed to start capture device\n");
-            return false;
-        }
-
-        return true;
-    }
-
-    bool startPlayback()
-    {
-        if (!playbackInitialized)
-        {
-            return false;
-        }
-
-        if (ma_device_start(&playbackDevice) != MA_SUCCESS)
-        {
-            printf("[!] Failed to start playback device\n");
-            return false;
-        }
-
-        return true;
-    }
-
-    bool restartPlayback()
-    {
-        if (!playbackInitialized)
-        {
-            return false;
-        }
-
-        ma_device_stop(&playbackDevice);
-
-        return ma_device_start(&playbackDevice) == MA_SUCCESS;
-    }
-
-    void stopDevices()
-    {
-        printf("[*] Stopping audio devices...\n");
-
-        if (captureInitialized)
-        {
-            ma_device_stop(&captureDevice);
-        }
-
-        if (playbackInitialized)
-        {
-            ma_device_stop(&playbackDevice);
-        }
-
-        // Give callbacks time to exit
-        ma_sleep(50);
-        printf("[+] Audio devices stopped\n");
-    }
-
-    // Forward declarations for callback functions
-    static void capture_callback(ma_device *device, void *output, const void *input, ma_uint32 frameCount);
-    static void playback_callback(ma_device *device, void *output, const void *input, ma_uint32 frameCount);
-
-  private:
-    ma_device captureDevice{};
-    ma_device playbackDevice{};
-    bool captureInitialized = false;
-    bool playbackInitialized = false;
-
-    void cleanup()
-    {
-        if (captureInitialized)
-        {
-            ma_device_uninit(&captureDevice);
-        }
-        if (playbackInitialized)
-        {
-            ma_device_uninit(&playbackDevice);
-        }
-    }
-
-};
 
 // === Improved Audio Callbacks ===
 void AudioDeviceManager::capture_callback(ma_device *device, void *output, const void *input, ma_uint32 frameCount)
